RE #27879 Fix interpolation setting not appearing in fig options
The setCurrentText method of the QComboBox is case sensitive, which is
incompatible as the values in the combo box are capitalised, and the
matplotlib values are not. Instead, a case insensitive search is done in
the list of combobox values to get the index and then the value of the
box is set to that when setting the value in the view, which happens
upon intitialisation.